+++ tests/docker_service_fixtures.py
@@ -43,7 +43,8 @@ class DockerServiceRegistry:
         self._running_services: set[str] = set()
         self.docker_ip = self._get_docker_ip()
         self._base_command = [
-            "docker-compose",
+            "docker",
+            "compose",
             "--file=tests/docker-compose.yml",
             "--project-name=litestar_users_pytest",
         ]